+++ oshi-core/src/main/java/oshi/hardware/platform/unix/solaris/SolarisFirmware.java
@@ -1,0 +1,23 @@
+package oshi.hardware.platform.unix.solaris;
+
+import oshi.hardware.common.AbstractFirmware;
+
+/**
+ * Created by IntelliJ IDEA.
+ * User: SchiTho1
+ * Date: 14.10.2016
+ * Time: 07:59
+ * <p>
+ * Copyright 2008-2013 - Securiton AG all rights reserved
+ */
+final class SolarisFirmware extends AbstractFirmware {
+
+    SolarisFirmware() {
+
+        init();
+    }
+
+    private void init() {
+        // TODO
+    }
+}