+++ oshi-core/src/main/java/oshi/software/os/windows/WindowsOSVersionInfoEx.java
@@ -19,21 +19,21 @@
 package oshi.software.os.windows;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.sun.jna.Native;
-import com.sun.jna.platform.win32.Advapi32Util;
-import com.sun.jna.platform.win32.Kernel32;
 import com.sun.jna.platform.win32.User32;
 import com.sun.jna.platform.win32.WinNT;
-import com.sun.jna.platform.win32.WinNT.OSVERSIONINFOEX;
-import com.sun.jna.platform.win32.WinReg;
 import com.sun.jna.platform.win32.WinUser;
 
 import oshi.software.common.AbstractOSVersionInfoEx;
+import oshi.util.ParseUtil;
+import oshi.util.platform.windows.WmiUtil;
+import oshi.util.platform.windows.WmiUtil.ValueType;
 
 public class WindowsOSVersionInfoEx extends AbstractOSVersionInfoEx {
 
@@ -41,35 +41,28 @@ public class WindowsOSVersionInfoEx extends AbstractOSVersionInfoEx {
 
     private static final Logger LOG = LoggerFactory.getLogger(WindowsOSVersionInfoEx.class);
 
-    private transient OSVERSIONINFOEX versionInfo = null;
+    private static final ValueType[] queryTypes = { ValueType.STRING, ValueType.UINT32, ValueType.STRING,
+            ValueType.STRING, ValueType.UINT32 };
+
+    private transient Map<String, List<Object>> versionInfo = new HashMap<>();
 
     public WindowsOSVersionInfoEx() {
-        this.versionInfo = new OSVERSIONINFOEX();
-        if (!Kernel32.INSTANCE.GetVersionEx(this.versionInfo)) {
-            LOG.error("Failed to Initialize OSVersionInfoEx. Error code: {}", Kernel32.INSTANCE.GetLastError());
-        }
-        init();
-        LOG.debug("Initialized OSVersionInfoEx");
-
-    }
-
-    public WindowsOSVersionInfoEx(OSVERSIONINFOEX versionInfo) {
-        this.versionInfo = versionInfo;
-        init();
-    }
-
-    /**
-     * Initialize class variables
-     */
-    private void init() {
-        setVersion(parseVersion());
-        setCodeName(parseCodeName());
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No build number available");
+        // Populate a key-value map from WMI
+        this.versionInfo = WmiUtil.selectObjectsFrom(null, "Win32_OperatingSystem",
+                "Version,ProductType,BuildNumber,CSDVersion,SuiteMask", null, queryTypes);
+        if (this.versionInfo.get("Version").isEmpty()) {
+            LOG.warn("No version data available.");
+            setVersion(System.getProperty("os.version"));
+            setCodeName("");
             setBuildNumber("");
-            return;
+        } else {
+            // Guaranteed that versionInfo is not null and lists non-empty
+            // before calling the parse*() methods
+            setVersion(parseVersion());
+            setCodeName(parseCodeName());
+            setBuildNumber(parseBuildNumber());
+            LOG.debug("Initialized OSVersionInfoEx");
         }
-        setBuildNumber(this.versionInfo.dwBuildNumber.toString());
     }
 
     /**
@@ -78,162 +71,56 @@ public class WindowsOSVersionInfoEx extends AbstractOSVersionInfoEx {
      * @return Version
      */
     private String parseVersion() {
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No version data available");
-            return System.getProperty("os.version");
-        }
 
-        String version;
+        // Initialize a default, sane value
+        String version = System.getProperty("os.version");
+
+        // Version is major.minor.build. Parse the version string for
+        // major/minor and get the build number separately
+        String[] verSplit = ((String) this.versionInfo.get("Version").get(0)).split("\\D");
+        int major = verSplit.length > 0 ? ParseUtil.parseIntOrDefault(verSplit[0], 0) : 0;
+        int minor = verSplit.length > 1 ? ParseUtil.parseIntOrDefault(verSplit[1], 0) : 0;
 
         // see
         // http://msdn.microsoft.com/en-us/library/windows/desktop/ms724833%28v=vs.85%29.aspx
-        if (getPlatformId() == WinNT.VER_PLATFORM_WIN32_NT) {
-            boolean ntWorkstation = getProductType() == WinNT.VER_NT_WORKSTATION;
-            if (getMajor() == 10 && getMinor() == 0 && ntWorkstation) {
-                version = "10";
-            } else if (getMajor() == 10 && getMinor() == 0 && !ntWorkstation) {
-                version = "Server 2016";
-            } else if (getMajor() == 6 && getMinor() == 3 && ntWorkstation) {
-                version = "8.1";
-            } else if (getMajor() == 6 && getMinor() == 3 && !ntWorkstation) {
-                version = "Server 2012 R2";
-            } else if (getMajor() == 6 && getMinor() == 2 && ntWorkstation) {
-                version = "8";
-            } else if (getMajor() == 6 && getMinor() == 2 && !ntWorkstation) {
-                version = "Server 2012";
-            } else if (getMajor() == 6 && getMinor() == 1 && ntWorkstation) {
-                version = "7";
-            } else if (getMajor() == 6 && getMinor() == 1 && !ntWorkstation) {
-                version = "Server 2008 R2";
-            } else if (getMajor() == 6 && getMinor() == 0 && !ntWorkstation) {
-                version = "Server 2008";
-            } else if (getMajor() == 6 && getMinor() == 0 && ntWorkstation) {
-                version = "Vista";
-            } else if (getMajor() == 5 && getMinor() == 2 && !ntWorkstation) {
-                version = User32.INSTANCE.GetSystemMetrics(WinUser.SM_SERVERR2) != 0 ? "Server 2003" : "Server 2003 R2";
-            } else if (getMajor() == 5 && getMinor() == 2 && ntWorkstation) {
-                version = "XP"; // 64 bits
-            } else if (getMajor() == 5 && getMinor() == 1) {
+        boolean ntWorkstation = (long) this.versionInfo.get("ProductType").get(0) == WinNT.VER_NT_WORKSTATION;
+        if (major == 10) {
+            if (minor == 0) {
+                version = ntWorkstation ? "10" : "Server 2016";
+            }
+        } else if (major == 6) {
+            if (minor == 3) {
+                version = ntWorkstation ? "8.1" : "Server 2012 R2";
+            } else if (minor == 2) {
+                version = ntWorkstation ? "8" : "Server 2012";
+            } else if (minor == 1) {
+                version = ntWorkstation ? "7" : "Server 2008 R2";
+            } else if (minor == 0) {
+                version = ntWorkstation ? "Vista" : "Server 2008";
+            }
+        } else if (major == 5) {
+            if (minor == 2) {
+                if ((getSuiteMask() & 0x00008000) != 0) {// VER_SUITE_WH_SERVER
+                    version = "Home Server";
+                } else if (ntWorkstation) {
+                    version = "XP"; // 64 bits
+                } else {
+                    version = User32.INSTANCE.GetSystemMetrics(WinUser.SM_SERVERR2) != 0 ? "Server 2003"
+                            : "Server 2003 R2";
+                }
+            } else if (minor == 1) {
                 version = "XP"; // 32 bits
-            } else if (getMajor() == 5 && getMinor() == 0) {
+            } else if (minor == 0) {
                 version = "2000";
-            } else if (getMajor() == 4) {
-                version = "NT 4";
-                if ("Service Pack 6".equals(getServicePack())
-                        && Advapi32Util.registryKeyExists(WinReg.HKEY_LOCAL_MACHINE,
-                                "SOFTWARE\\Microsoft\\Windows NT\\CurrentVersion\\Hotfix\\Q246009")) {
-                    return "NT4 SP6a";
-                }
-            } else {
-                throw new UnsupportedOperationException(
-                        "Unsupported Windows NT version: " + this.versionInfo.toString());
             }
+        }
 
-            if (this.versionInfo.wServicePackMajor.intValue() > 0) {
-                version = version + " SP" + this.versionInfo.wServicePackMajor.intValue();
-            }
-
-        } else if (getPlatformId() == WinNT.VER_PLATFORM_WIN32_WINDOWS) {
-            if (getMajor() == 4 && getMinor() == 90) {
-                version = "ME";
-            } else if (getMajor() == 4 && getMinor() == 10) {
-                if (this.versionInfo.szCSDVersion[1] == 'A') {
-                    version = "98 SE";
-                } else {
-                    version = "98";
-                }
-            } else if (getMajor() == 4 && getMinor() == 0) {
-                if (this.versionInfo.szCSDVersion[1] == 'C' || this.versionInfo.szCSDVersion[1] == 'B') {
-                    version = "95 OSR2";
-                } else {
-                    version = "95";
-                }
-            } else {
-                throw new UnsupportedOperationException(
-                        "Unsupported Windows 9x version: " + this.versionInfo.toString());
-            }
-        } else {
-            throw new UnsupportedOperationException("Unsupported Windows platform: " + this.versionInfo.toString());
+        String sp = (String) this.versionInfo.get("CSDVersion").get(0);
+        if (!sp.isEmpty() && !"unknown".equals(sp)) {
+            version = version + " " + sp.replace("Service Pack ", "SP");
         }
 
         return version;
-    }
-
-    /**
-     * The operating system platform. This member can be VER_PLATFORM_WIN32_NT.
-     *
-     * @return Platform ID.
-     */
-    private int getPlatformId() {
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No platform id available");
-            return 0;
-        }
-        return this.versionInfo.dwPlatformId.intValue();
-    }
-
-    /**
-     * Any additional information about the system.
-     *
-     * @return Product type.
-     */
-    private byte getProductType() {
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No product type available");
-            return 0;
-        }
-        return this.versionInfo.wProductType;
-    }
-
-    /**
-     * The major version number of the operating system.
-     *
-     * @return The major version within the following supported operating
-     *         systems. Windows 8: 6.2 Windows Server 2012: 6.2 Windows 7: 6.1
-     *         Windows Server 2008 R2: 6.1 Windows Server 2008: 6.0 Windows
-     *         Vista: 6.0 Windows Server 2003 R2: 5.2 Windows Home Server: 5.2
-     *         Windows Server 2003: 5.2 Windows XP Professional x64 Edition: 5.2
-     *         Windows XP: 5.1 Windows 2000: 5.0
-     */
-    private int getMajor() {
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No version data available");
-            return 0;
-        }
-        return this.versionInfo.dwMajorVersion.intValue();
-    }
-
-    /**
-     * The minor version number of the operating system.
-     *
-     * @return The minor version within the following supported operating
-     *         systems. Windows 8: 6.2 Windows Server 2012: 6.2 Windows 7: 6.1
-     *         Windows Server 2008 R2: 6.1 Windows Server 2008: 6.0 Windows
-     *         Vista: 6.0 Windows Server 2003 R2: 5.2 Windows Home Server: 5.2
-     *         Windows Server 2003: 5.2 Windows XP Professional x64 Edition: 5.2
-     *         Windows XP: 5.1 Windows 2000: 5.0
-     */
-    private int getMinor() {
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No version data available");
-            return 0;
-        }
-        return this.versionInfo.dwMinorVersion.intValue();
-    }
-
-    /**
-     * String, such as "Service Pack 3", that indicates the latest Service Pack
-     * installed on the system. If no Service Pack has been installed, the
-     * string is empty.
-     *
-     * @return Service pack.
-     */
-    private String getServicePack() {
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No service pack data available");
-            return "";
-        }
-        return Native.toString(this.versionInfo.szCSDVersion);
     }
 
     /**
@@ -244,27 +131,31 @@ public class WindowsOSVersionInfoEx extends AbstractOSVersionInfoEx {
     private String parseCodeName() {
         List<String> suites = new ArrayList<>();
         int bitmask = getSuiteMask();
+        if ((bitmask & 0x00000002) != 0) {
+            suites.add("Enterprise");
+        }
         if ((bitmask & 0x00000004) != 0) {
             suites.add("BackOffice");
         }
-        if ((bitmask & 0x00000400) != 0) {
-            suites.add("Web Edition");
-        }
-        if ((bitmask & 0x00004000) != 0) {
-            suites.add("Compute Cluster");
+        if ((bitmask & 0x00000008) != 0) {
+            suites.add("Communication Server");
         }
         if ((bitmask & 0x00000080) != 0) {
             suites.add("Datacenter");
         }
-        if ((bitmask & 0x00000002) != 0) {
-            suites.add("Enterprise");
-        }
         if ((bitmask & 0x00000200) != 0) {
             suites.add("Home");
         }
-        if ((bitmask & 0x00008000) != 0) {
-            suites.add("Home Server");
+        if ((bitmask & 0x00000400) != 0) {
+            suites.add("Web Server");
         }
+        if ((bitmask & 0x00002000) != 0) {
+            suites.add("Storage Server");
+        }
+        if ((bitmask & 0x00004000) != 0) {
+            suites.add("Compute Cluster");
+        }
+        // 0x8000, Home Server, is included in main version name
         String separator = "";
         StringBuilder sb = new StringBuilder();
         for (String s : suites) {
@@ -280,11 +171,19 @@ public class WindowsOSVersionInfoEx extends AbstractOSVersionInfoEx {
      * @return Suite mask.
      */
     private int getSuiteMask() {
-        if (this.versionInfo == null) {
-            LOG.warn("OSVersionInfoEx not initialized. No suite mask data available");
-            return 0;
-        }
-
-        return this.versionInfo.wSuiteMask.intValue();
+        // Although this object is 64 bits, it originates from
+        // a UINT32 bit mask and can safely be cast directly
+        // to int, which preserves the low 32 bits
+        return (int) ((Long) this.versionInfo.get("SuiteMask").get(0)).longValue();
     }
+
+    /**
+     * Gets the build number
+     * 
+     * @return A string representing the Build Number
+     */
+    private String parseBuildNumber() {
+        return (String) this.versionInfo.get("BuildNumber").get(0);
+    }
+
 }