+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 3.3 (in progress)
 ================
-* Your contribution here.
-* [#283](https://github.com/dblock/oshi/pull/283): Bugfix for Linux block device stats - [@lu-ko](https://github.com/lu-ko).
 * [#262](https://github.com/dblock/oshi/pull/262): Add bytesRead and bytesWritten to OSProcess - [@plamenko](https://github.com/plamenko).
+* [#264](https://github.com/dblock/oshi/pull/264): BIOS and manufacturer information - [@lundefugl](https://github.com/lundefugl).
+* [#281](https://github.com/dblock/oshi/pull/281): Improve Linux battery AC device exclusion - [@dbwiddis](https://github.com/dbwiddis).
+* [#282](https://github.com/dblock/oshi/pull/282): Get Windows version from WMI - [@dbwiddis](https://github.com/dbwiddis).
+* [#283](https://github.com/dblock/oshi/pull/283): Bugfix for Linux block device stats - [@lu-ko](https://github.com/lu-ko).
+* Your contribution here.
 
 3.2 (9/1/2016)
 ================